ms-inventario/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
SQLAlchemy==1.4.46
Werkzeug==2.0.3
python-dotenv==1.0.0